templates/requirements.txt
# For scripts/stitch.py (used by release workflow)
pyyaml>=6.0
